xml/parsing/manakai/templates-1.dat

#data
<template>aa</template>
#errors
1;1;no XML decl;;;s
#document
| <{} template>
|   "aa"

#data
<template xmlns="http://www.w3.org/1999/xhtml">aa</template>
#errors
1;1;no XML decl;;;s
#document
| <template>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   content
|     "aa"

#data
<template xmlns="http://www.w3.org/1999/xhtml"><template>aa</template></template>
#errors
1;1;no XML decl;;;s
#document
| <template>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   content
|     <template>
|       content
|         "aa"
